dashboard/requirements.txt
requests==2.31.0
PyGithub==2.1.1
google-generativeai==0.3.2
supabase==2.3.0
pytz==2023.3

